add config option PRINTF_PROVIDE_PLAIN_PRINTF When disabled, the main header doesn't expose printf_ and vprintf_, and these functins are not compiled into the library. Code is split into separate translation units for this purpose, and to allow linking with the full (static) library without supplying a definition for putchar_.
