+++ printf_config.h.in
@@ -17,6 +17,7 @@
 #define PRINTF_MAX_INTEGRAL_DIGITS_FOR_DECIMAL  @MAX_INTEGRAL_DIGITS_FOR_DECIMAL@
 #define PRINTF_LOG10_TAYLOR_TERMS               @LOG10_TAYLOR_TERMS@
 #define PRINTF_CHECK_FOR_NUL_IN_FORMAT_SPECIFIER @PRINTF_CHECK_FOR_NUL_IN_FORMAT_SPECIFIER@
+#define PRINTF_PROVIDE_PLAIN_PRINTF             @PRINTF_PROVIDE_PLAIN_PRINTF@
 
 #endif /* PRINTF_CONFIG_H_ */
