+++ test/test_suite.cpp
@@ -32,6 +32,7 @@
 #if PRINTF_INCLUDE_CONFIG_H
 #include <printf_config.h>
 #endif
+#include <printf/printf_core.c>
 #include <printf/printf.c>
 
 #if PRINTF_ALIAS_STANDARD_FUNCTION_NAMES_HARD
